8949: Move allow origin header handling into CORS middleware
